Module aws_sdk_workmailmessageflow::error

Modules

invalid_content_location

See InvalidContentLocation

message_frozen

See MessageFrozen

message_rejected

See MessageRejected

resource_not_found_exception

See ResourceNotFoundException

Structs

GetRawMessageContentError

InvalidContentLocation

WorkMail could not access the updated email content. Possible reasons:

MessageFrozen

The requested email is not eligible for update. This is usually the case for a redirected email.

MessageRejected

The requested email could not be updated due to an error in the MIME content. Check the error message for more information about

PutRawMessageContentError

ResourceNotFoundException

The requested email message is not found.

Enums

GetRawMessageContentErrorKind

PutRawMessageContentErrorKind